@@ -21,16 +21,16 @@
     "eslint-config-next": "13.3.1",
     "immer": "^10.0.1",
     "js-cookie": "^3.0.5",
-    "next": "13.3.1",
-    "next-auth": "^4.22.1",
+    "next": "13.4.6",
     "notistack": "^3.0.1",
     "react": "18.2.0",
     "react-datepicker": "^4.11.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.43.9",
-    "react-query": "^3.39.3",
+    "server-only": "^0.0.1",
     "typescript": "5.0.4",
-    "use-immer": "^0.9.0"
+    "use-immer": "^0.9.0",
+    "zod": "^3.21.4"
   },
   "devDependencies": {
     "@svgr/webpack": "^7.0.0",
