@@ -1,5 +1,7 @@
+"use client";
+
 import React from "react";
-import { useRouter } from "next/router";
+import { usePathname } from "next/navigation";
 import Link, { LinkProps } from "next/link";
 
 export interface NavLinkProps extends 
@@ -17,10 +19,21 @@ export interface NavLinkProps extends
 }
 
 export const NavLink: React.FC<NavLinkProps> = ({ classes, children, ...linkProps }) => {
-  const router = useRouter();
+  const pathname = usePathname();
 
   const isLinkActive = () => {
-    return router.pathname === linkProps.href;
+    if (!linkProps.href) return false;
+    
+    let href = "";
+    if (typeof linkProps.href === "string") {
+      href = linkProps.href;
+    }
+    else {
+      if (!linkProps.href.href) return false;
+      href = linkProps.href.href;
+    }
+
+    return pathname?.startsWith(href);
   };
 
   return (
