@@ -1,9 +1,13 @@
-import os
+import jwt
+from django.conf import settings
 from django.contrib.sites.shortcuts import get_current_site
 from django.core.mail import EmailMessage, EmailMultiAlternatives
 from django.template import TemplateDoesNotExist
 from django.template.loader import render_to_string
 from allauth.account.adapter import DefaultAccountAdapter
+from allauth.socialaccount.adapter import app_settings, DefaultSocialAccountAdapter
+from allauth.socialaccount.providers.oauth2.client import OAuth2Error
+from allauth.socialaccount.providers.google.views import GoogleOAuth2Adapter as DefaultGoogleOAuth2Adapter
 
 from .functions import get_frontend_url
 
@@ -12,7 +16,7 @@ class AccountAdapter(DefaultAccountAdapter):
     def send_confirmation_mail(self, request, emailconfirmation, signup):
         frontend_url = get_frontend_url()
         # activation_url = f"{frontend_url}/auth/register/{emailconfirmation.key}"
-        register_confirm_path = os.getenv("FRONTEND_REGISTER_CONFIRM_PATH", "")
+        register_confirm_path = settings.ENV["FRONTEND_REGISTER_CONFIRM_PATH"]
         register_confirm_path = register_confirm_path.replace("<key>", emailconfirmation.key)
         activation_url = f"{frontend_url}/{register_confirm_path}"
         
@@ -62,3 +66,47 @@ class AccountAdapter(DefaultAccountAdapter):
             msg = EmailMessage(subject, bodies["html"], from_email, to, headers=headers)
             msg.content_subtype = "html"  # Main content is now text/html
         return msg
+
+
+class SocialAccountAdapter(DefaultSocialAccountAdapter):
+    def get_app(self, request, provider, config=None):
+        # NOTE: Avoid loading models at top due to registry boot...
+        from allauth.socialaccount.models import SocialApp
+
+        config = config or app_settings.PROVIDERS.get(provider, {}).get("APP")
+        if config:
+            # app = SocialApp(provider=provider)
+            app = SocialApp.objects.get_or_create(provider=provider)[0]
+            for field in ["client_id", "secret", "key", "certificate_key"]:
+                setattr(app, field, config.get(field))
+        else:
+            app = SocialApp.objects.get_current(provider, request)
+
+        return app
+
+
+class GoogleOAuth2Adapter(DefaultGoogleOAuth2Adapter):
+    
+     def complete_login(self, request, app, token, response, **kwargs):
+        try:
+            identity_data = jwt.decode(
+                response["id_token"]["id_token"],
+                # Since the token was received by direct communication
+                # protected by TLS between this library and Google, we
+                # are allowed to skip checking the token signature
+                # according to the OpenID Connect Core 1.0
+                # specification.
+                # https://openid.net/specs/openid-connect-core-1_0.html#IDTokenValidation
+                options={
+                    "verify_signature": False,
+                    "verify_iss": True,
+                    "verify_aud": True,
+                    "verify_exp": True,
+                },
+                issuer=self.id_token_issuer,
+                audience=app.client_id,
+            )
+        except jwt.PyJWTError as e:
+            raise OAuth2Error("Invalid id_token") from e
+        login = self.get_provider().sociallogin_from_response(request, identity_data)
+        return login
